Update test directory to include unit and integration directories; moved files not intended for automated tests to root of test
