+++ tests/slackBotTest.js
