+++ tests/unit/models/scheduledTripsModel.test.js
